menu de pausa con la vista del deck y añadido dos texts a las cartas para coste y daño

+++ Assets/Scripts/Cards/CardData.cs
@@ -6,7 +6,7 @@ public abstract class CardData : ScriptableObject
 {
     [Header("Datos comunes")]
     public string cardName;
-    public byte cost;
+    public float cost;
     [TextArea] public string description;
     public CardType cardType;
